Add ConnectionConfig, rename ConnectConfig to SessionConfig
Session is now configured using SessionConfig which contains
list of known nodes and some connection options. It can be converted
to ConnectionConfig which is passed to Connection with options
like Compression etc.
